make shortcut expanding restmapper handle all
